Bench 3084378

+++ src/search.h
@@ -39,7 +39,7 @@ namespace Search {
     int doubleExt;
 
     // [piece to]
-    int* contHistory;
+    int16_t* contHistory;
   };
 
   // A sort of header of the search stack, so that plies behind 0 are accessible and
